#126 자신이 작성한 리뷰 및 포스트에는 좋아요 누르지 못하도록 처리

+++ api/src/main/java/com/hot6/phopa/api/domain/community/service/PostApiService.java
@@ -140,8 +140,11 @@ public class PostApiService {
 
     public void like(Long postId) {
         UserDTO userDTO = PrincipleDetail.get();
-        UserEntity userEntity = userDTO.getId() != null ? userService.findById(userDTO.getId()) : null;
+        UserEntity userEntity = userService.findById(userDTO.getId());
         PostEntity postEntity = postService.getPostById(postId);
+        if(userEntity.getId().equals(postEntity.getUser().getId())){
+            throw new SilentApplicationErrorException(ApplicationErrorType.CANNOT_BE_CREATED_USER);
+        }
         PostLikeEntity postLikeEntity = postService.getPostLikeByPostIdAndUserId(postId, userEntity.getId());
         if (postLikeEntity != null) {
             postService.deletePostLikeEntity(postLikeEntity);
